fix: fetching processing
Fix fetching models processing with a better way to stop it.
Also remove setting the tab when opening window and only let it inside the OnEnable. Avoid double fetching.

With this fix we also fix another bug, which it was, sometines on redraw or reopening first models or the first page was in double.